+++ journal/src/main/java/io/zeebe/journal/file/record/JournalRecordMetadata.java
@@ -19,5 +19,5 @@ public interface JournalRecordMetadata {
 
   long checksum();
 
-  long length();
+  int length();
 }
