+++ journal/src/main/java/io/zeebe/journal/file/record/JournalRecordMetadataImpl.java
@@ -43,7 +43,7 @@ public final class JournalRecordMetadataImpl implements JournalRecordMetadata {
   }
 
   @Override
-  public long length() {
+  public int length() {
     return length;
   }
